Add files via upload
At the end I found that adding a new file directly is the easiest way to push the change haha
Just need to drop the old notebook once the new notebook is finalized.

Repeated my email here:

We're almost done! I've finished most of the discussion questions and fixed some of the plots. The only things left are the following:
1. Explain subgroup distribution plot
2. Add description of Elastic-Net
3. Add description of Random Forest
4. (Optional) Make the correlation plot look better. It already looks much better now, but I still think the small plots aren't clearly separated from each other. If you have time (and the energy) after everything else is done, feel free to take a look. Otherwise we can leave the plot as is.

And of course feel free to change any part of the discussion if you think there are any issues. After that we're good to go!